// web/src/hooks/queries/gql-generated-uniswap.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  BigDecimal: { input: string; output: string };
  BigInt: { input: string; output: string };
  Bytes: { input: `0x${string}`; output: `0x${string}` };
  Int8: { input: string; output: string };
  Timestamp: { input: string; output: string };
};
// ...
export type Pool = {
  __typename?: "Pool";
  burns: Array<Burn>;
  collectedFeesToken0: Scalars["BigDecimal"]["output"];
  collectedFeesToken1: Scalars["BigDecimal"]["output"];
  collectedFeesUSD: Scalars["BigDecimal"]["output"];
  collects: Array<Collect>;
  createdAtBlockNumber: Scalars["BigInt"]["output"];
  createdAtTimestamp: Scalars["BigInt"]["output"];
  feeGrowthGlobal0X128: Scalars["BigInt"]["output"];
  feeGrowthGlobal1X128: Scalars["BigInt"]["output"];
  feeTier: Scalars["BigInt"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  liquidity: Scalars["BigInt"]["output"];
  liquidityProviderCount: Scalars["BigInt"]["output"];
  mints: Array<Mint>;
  observationIndex: Scalars["BigInt"]["output"];
  poolDayData: Array<PoolDayData>;
  poolHourData: Array<PoolHourData>;
  sqrtPrice: Scalars["BigInt"]["output"];
  swaps: Array<Swap>;
  tick?: Maybe<Scalars["BigInt"]["output"]>;
  ticks: Array<Tick>;
  token0: Token;
  token0Price: Scalars["BigDecimal"]["output"];
  token1: Token;
  token1Price: Scalars["BigDecimal"]["output"];
  totalValueLockedETH: Scalars["BigDecimal"]["output"];
  totalValueLockedToken0: Scalars["BigDecimal"]["output"];
  totalValueLockedToken1: Scalars["BigDecimal"]["output"];
  totalValueLockedUSD: Scalars["BigDecimal"]["output"];
  totalValueLockedUSDUntracked: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  untrackedVolumeUSD: Scalars["BigDecimal"]["output"];
  volumeToken0: Scalars["BigDecimal"]["output"];
  volumeToken1: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export type PoolDayData = {
  __typename?: "PoolDayData";
  close: Scalars["BigDecimal"]["output"];
  date: Scalars["Int"]["output"];
  feeGrowthGlobal0X128: Scalars["BigInt"]["output"];
  feeGrowthGlobal1X128: Scalars["BigInt"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  high: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  liquidity: Scalars["BigInt"]["output"];
  low: Scalars["BigDecimal"]["output"];
  open: Scalars["BigDecimal"]["output"];
  pool: Pool;
  sqrtPrice: Scalars["BigInt"]["output"];
  tick?: Maybe<Scalars["BigInt"]["output"]>;
  token0Price: Scalars["BigDecimal"]["output"];
  token1Price: Scalars["BigDecimal"]["output"];
  tvlUSD: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  volumeToken0: Scalars["BigDecimal"]["output"];
  volumeToken1: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export type PoolHourData = {
  __typename?: "PoolHourData";
  close: Scalars["BigDecimal"]["output"];
  feeGrowthGlobal0X128: Scalars["BigInt"]["output"];
  feeGrowthGlobal1X128: Scalars["BigInt"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  high: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  liquidity: Scalars["BigInt"]["output"];
  low: Scalars["BigDecimal"]["output"];
  open: Scalars["BigDecimal"]["output"];
  periodStartUnix: Scalars["Int"]["output"];
  pool: Pool;
  sqrtPrice: Scalars["BigInt"]["output"];
  tick?: Maybe<Scalars["BigInt"]["output"]>;
  token0Price: Scalars["BigDecimal"]["output"];
  token1Price: Scalars["BigDecimal"]["output"];
  tvlUSD: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  volumeToken0: Scalars["BigDecimal"]["output"];
  volumeToken1: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export enum Pool_OrderBy {
  Burns = "burns",
  CollectedFeesToken0 = "collectedFeesToken0",
  CollectedFeesToken1 = "collectedFeesToken1",
  CollectedFeesUsd = "collectedFeesUSD",
  Collects = "collects",
  CreatedAtBlockNumber = "createdAtBlockNumber",
  CreatedAtTimestamp = "createdAtTimestamp",
  FeeGrowthGlobal0X128 = "feeGrowthGlobal0X128",
  FeeGrowthGlobal1X128 = "feeGrowthGlobal1X128",
  FeeTier = "feeTier",
  FeesUsd = "feesUSD",
  Id = "id",
  Liquidity = "liquidity",
  LiquidityProviderCount = "liquidityProviderCount",
  Mints = "mints",
  ObservationIndex = "observationIndex",
  PoolDayData = "poolDayData",
  PoolHourData = "poolHourData",
  SqrtPrice = "sqrtPrice",
  Swaps = "swaps",
  Tick = "tick",
  Ticks = "ticks",
  Token0 = "token0",
  Token0Price = "token0Price",
  Token0Decimals = "token0__decimals",
  Token0DerivedEth = "token0__derivedETH",
  Token0FeesUsd = "token0__feesUSD",
  Token0Id = "token0__id",
  Token0Name = "token0__name",
  Token0PoolCount = "token0__poolCount",
  Token0Symbol = "token0__symbol",
  Token0TotalSupply = "token0__totalSupply",
  Token0TotalValueLocked = "token0__totalValueLocked",
  Token0TotalValueLockedUsd = "token0__totalValueLockedUSD",
  Token0TotalValueLockedUsdUntracked = "token0__totalValueLockedUSDUntracked",
  Token0TxCount = "token0__txCount",
  Token0UntrackedVolumeUsd = "token0__untrackedVolumeUSD",
  Token0Volume = "token0__volume",
  Token0VolumeUsd = "token0__volumeUSD",
  Token1 = "token1",
  Token1Price = "token1Price",
  Token1Decimals = "token1__decimals",
  Token1DerivedEth = "token1__derivedETH",
  Token1FeesUsd = "token1__feesUSD",
  Token1Id = "token1__id",
  Token1Name = "token1__name",
  Token1PoolCount = "token1__poolCount",
  Token1Symbol = "token1__symbol",
  Token1TotalSupply = "token1__totalSupply",
  Token1TotalValueLocked = "token1__totalValueLocked",
  Token1TotalValueLockedUsd = "token1__totalValueLockedUSD",
  Token1TotalValueLockedUsdUntracked = "token1__totalValueLockedUSDUntracked",
  Token1TxCount = "token1__txCount",
  Token1UntrackedVolumeUsd = "token1__untrackedVolumeUSD",
  Token1Volume = "token1__volume",
  Token1VolumeUsd = "token1__volumeUSD",
  TotalValueLockedEth = "totalValueLockedETH",
  TotalValueLockedToken0 = "totalValueLockedToken0",
  TotalValueLockedToken1 = "totalValueLockedToken1",
  TotalValueLockedUsd = "totalValueLockedUSD",
  TotalValueLockedUsdUntracked = "totalValueLockedUSDUntracked",
  TxCount = "txCount",
  UntrackedVolumeUsd = "untrackedVolumeUSD",
  VolumeToken0 = "volumeToken0",
  VolumeToken1 = "volumeToken1",
  VolumeUsd = "volumeUSD",
}
// ...
export type Tick = {
  __typename?: "Tick";
  collectedFeesToken0: Scalars["BigDecimal"]["output"];
  collectedFeesToken1: Scalars["BigDecimal"]["output"];
  collectedFeesUSD: Scalars["BigDecimal"]["output"];
  createdAtBlockNumber: Scalars["BigInt"]["output"];
  createdAtTimestamp: Scalars["BigInt"]["output"];
  feeGrowthOutside0X128: Scalars["BigInt"]["output"];
  feeGrowthOutside1X128: Scalars["BigInt"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  liquidityGross: Scalars["BigInt"]["output"];
  liquidityNet: Scalars["BigInt"]["output"];
  liquidityProviderCount: Scalars["BigInt"]["output"];
  pool: Pool;
  poolAddress?: Maybe<Scalars["String"]["output"]>;
  price0: Scalars["BigDecimal"]["output"];
  price1: Scalars["BigDecimal"]["output"];
  tickIdx: Scalars["BigInt"]["output"];
  untrackedVolumeUSD: Scalars["BigDecimal"]["output"];
  volumeToken0: Scalars["BigDecimal"]["output"];
  volumeToken1: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
};